Split tx_pool and tx_list unit and benchmark tests into separate files
